change: Rename port list component js module
We have several Component modules therefore use PortListComponent as the
js module name to be more explicit.
