+++ src/web/pages/portlists/ListPage.jsx
@@ -31,8 +31,8 @@ import useSelection from 'web/hooks/useSelection';
 import useShallowEqualSelector from 'web/hooks/useShallowEqualSelector';
 import useTranslation from 'web/hooks/useTranslation';
 import useUserSessionTimeout from 'web/hooks/useUserSessionTimeout';
-import PortListComponent from 'web/pages/portlists/Component';
 import PortListsFilterDialog from 'web/pages/portlists/FilterDialog';
+import PortListComponent from 'web/pages/portlists/PortListComponent';
 import PortListsTable from 'web/pages/portlists/Table';
 import {loadEntities, selector} from 'web/store/entities/portlists';
 import {getUserSettingsDefaults} from 'web/store/usersettings/defaults/selectors';
